pindah fungsi check balance ke database
