+++ database/database_virtual_accounts.go
@@ -15,3 +15,27 @@ func GetBalanceVA(vaNum string, db *sql.DB) (va models.VirtualAccounts, err erro
 	va.VaBalance = int(balanceFloat)
 	return
 }
+
+func CheckBalance(target string, accNumber string, amount int, db *sql.DB) (status bool) {
+	if target == "MAIN" {
+		sourceBalance, err := GetBalanceAcc(accNumber, db)
+		if err != nil {
+			return
+		}
+		if sourceBalance.AccountBalance < amount || amount <= 0 {
+			return
+		}
+		status = true
+	}
+	if target == "VA" {
+		sourceBalance, err := GetBalanceVA(accNumber, db)
+		if err != nil {
+			return
+		}
+		if sourceBalance.VaBalance < amount || amount <= 0 {
+			return
+		}
+		status = true
+	}
+	return
+}
